new update

+++ ft_libprintf.h
@@ -20,7 +20,7 @@
 int		ft_putuns(unsigned int n);
 int		ft_printf(char *str, ...);
 char	*ft_utoa(unsigned int n);
-int		ft_puthex(unsigned long n, char *f);
+int		ft_puthex(unsigned int n, char *f);
 int		ft_putptr(void *ptr);
 int		ft_putnbr(int n);
 int		ft_putchar(char c);
